MPT-17825: Split audit_records helpers into audit_paths module
Refactor the audit plugin so it no longer trips WPS210 (too many
locals), WPS231 (cognitive complexity) or WPS202 (too many module
members):

- New module `cli/plugins/audit_plugin/audit_paths.py` hosts the
  path/trail utilities (`AuditTrail`, `is_valid_path`,
  `get_external_id`, `parse_array_path`, `walk_trail`).
- `flatten_dict` keeps its recursive shape but the list-handling branch
  now delegates to a protected `_flatten_list_entries` helper that
  recurses back into `flatten_dict` for nested dicts. This collapses
  the inner loop/nesting that drove the WPS231 violation.
- `format_json_path` becomes a flat dispatch: parse the array path
  once, walk each trail, and ask `get_external_id` for the lookup.
  No more inline path parsing or trail walking inside the function;
  WPS210 falls below threshold without forcing a `# noqa`.
- Tests import `get_external_id` / `is_valid_path` from the new
  `audit_paths` module; behaviour is identical.

Drop the `WPS210,WPS231` entry for `audit_records.py` from
`[tool.flake8] per-file-ignores`. No behavior changes; ruff / flake8 /
mypy / pytest all green (363 passed, 100% coverage on both refactored
modules).
